Add autopep8.enabled to the schema
Added a note that in order to enable yapf, one must disable autopep8.

Confirmed with the following config, which reports to the log which
formatter is being used when invoking
`<cmd>lua vim.lsp.buf.formatting()<CR>`

```
nvim_lsp.pylsp.setup {
  cmd = { "/.../bin/pylsp", "-v", "--log-file=/tmp/pylsp.log" },
  settings = {
    pylsp = {
      plugins = {
        autopep8 = {enabled = false},
      },
    },
  },
}
```

I only noticed this because my yapf config uses 2-space indenting
(thanks, Google, for forcing me to do this), and with no configuration
changes it will ignore that config in my `setup.cfg`, because autopep8
doesn't see it.