Refine diagnostic severity for flake8
Previously all diagnostics that starts with "F" (generated from pyflakes)
are reported as errors, which could be too high. As such, in this commit
we change to report only certain pyflakes diagnostics as errors and all
the others as warnings. The current list of error codes is in sync with
the `pyflakes_lint` plugin.

While it is possible to generate such a list (or more precisely, a `set`)
during runtime (see the snippet in the comment), that might be less ideal
because:

  - Flake8 does not guarantee a stable API.
  - We run `flake8` as an executable and parse its outputs, so using the
    `flake8` module in Python may bring inconsistency (e.g., the user
    install Flake8 in a different virtual environment from pylsp).
  - Doing so couples two plugins (flake8 and pyflakes) together.

So it seems better to maintain such a list of error codes manually.

Fix #256.